+++ src/acodex/core/asyncio/cdp/__init__.py
@@ -1,14 +1,14 @@
 from __future__ import annotations
 
-from acodex.asyncio.cdp.client import CodexAppCdpClient
-from acodex.asyncio.cdp.errors import (
+from acodex.core.asyncio.cdp.client import CodexAppCdpClient
+from acodex.core.asyncio.cdp.errors import (
     CodexAppCdpConnectionError,
     CodexAppCdpDiscoveryError,
     CodexAppCdpError,
     CodexAppCdpEvaluationError,
     CodexAppCdpProtocolError,
 )
-from acodex.asyncio.cdp.renderer import (
+from acodex.core.asyncio.cdp.renderer import (
     ALL_CODEX_APP_THREAD_TOOL_NAMES,
     MUTATING_CODEX_APP_THREAD_TOOL_NAMES,
     READ_ONLY_CODEX_APP_THREAD_TOOL_NAMES,
@@ -16,8 +16,8 @@ from acodex.asyncio.cdp.renderer import (
     build_tool_invocation_expression,
     parse_tool_discovery_result,
 )
-from acodex.asyncio.cdp.runtime import parse_runtime_evaluate_response
-from acodex.asyncio.cdp.settings import (
+from acodex.core.asyncio.cdp.runtime import parse_runtime_evaluate_response
+from acodex.core.asyncio.cdp.settings import (
     DEFAULT_CDP_ENDPOINT,
     DEFAULT_CDP_HTTP_TIMEOUT,
     DEFAULT_CDP_RUNTIME_TIMEOUT,
@@ -25,12 +25,12 @@ from acodex.asyncio.cdp.settings import (
     DEFAULT_CDP_TARGET_URL_PREFIX,
     CodexAppCdpSettings,
 )
-from acodex.asyncio.cdp.targets import (
+from acodex.core.asyncio.cdp.targets import (
     fetch_cdp_targets,
     parse_cdp_targets,
     select_codex_app_target,
 )
-from acodex.asyncio.cdp.types import CdpTarget, CodexAppToolDiscovery, JsonObject, JsonValue
+from acodex.core.asyncio.cdp.types import CdpTarget, CodexAppToolDiscovery, JsonObject, JsonValue
 
 __all__ = [
     "ALL_CODEX_APP_THREAD_TOOL_NAMES",
