+++ src/acodex/core/asyncio/tools/fork_thread.py
@@ -5,7 +5,7 @@ from typing import Annotated, Any, ClassVar, Literal, TypeAlias
 from pydantic import Field, model_validator
 from typing_extensions import NotRequired, TypedDict, Unpack
 
-from acodex.asyncio.tools.base import BaseAsyncTool, RendererToolOutput
+from acodex.core.asyncio.tools.base import BaseAsyncTool, RendererToolOutput
 
 # Environment schemas are open-ended renderer payloads; JSON validation happens after dump.
 RendererJsonObject: TypeAlias = dict[str, Any]
