+++ src/acodex/core/asyncio/tools/handoff_thread.py
@@ -5,7 +5,7 @@ from typing import Annotated, ClassVar
 from pydantic import Field
 from typing_extensions import NotRequired, TypedDict, Unpack
 
-from acodex.asyncio.tools.base import BaseAsyncTool, RendererToolOutput
+from acodex.core.asyncio.tools.base import BaseAsyncTool, RendererToolOutput
 
 
 class HandoffThreadToolInput(TypedDict):
