+++ src/acodex/core/asyncio/tools/read_thread.py
@@ -5,8 +5,8 @@ from typing import Annotated, Any, ClassVar, Literal, TypeAlias
 from pydantic import Field
 from typing_extensions import NotRequired, TypedDict, Unpack
 
-from acodex.asyncio.cdp.types import ThinkingEffort
-from acodex.asyncio.tools.base import BaseAsyncTool, RendererToolOutput
+from acodex.core.asyncio.cdp.types import ThinkingEffort
+from acodex.core.asyncio.tools.base import BaseAsyncTool, RendererToolOutput
 
 # Some read_thread item fields contain arbitrary renderer JSON; envelope validation handles JSON.
 ReadThreadJsonValue: TypeAlias = Any
