+++ tests/tools/test_thread_tools.py
@@ -11,54 +11,54 @@ from pydantic import ValidationError
 from typing_extensions import Unpack
 
 from acodex import CodexAppCdpProtocolError, JsonObject, JsonValue
-from acodex.asyncio.tools import base as tools_base
-from acodex.asyncio.tools.base import dump_tool_input, parse_tool_output
-from acodex.asyncio.tools.create_thread import (
+from acodex.core.asyncio.tools import base as tools_base
+from acodex.core.asyncio.tools.base import dump_tool_input, parse_tool_output
+from acodex.core.asyncio.tools.create_thread import (
     CreateThreadTool,
     CreateThreadToolInput,
     CreateThreadToolOutput,
 )
-from acodex.asyncio.tools.fork_thread import (
+from acodex.core.asyncio.tools import (
     ForkThreadTool,
     ForkThreadToolInput,
     ForkThreadToolOutput,
 )
-from acodex.asyncio.tools.handoff_thread import (
+from acodex.core.asyncio.tools.handoff_thread import (
     HandoffThreadTool,
     HandoffThreadToolInput,
     HandoffThreadToolOutput,
 )
-from acodex.asyncio.tools.list_threads import (
+from acodex.core.asyncio.tools import (
     ListThreadsTool,
     ListThreadsToolInput,
     ListThreadsToolOutput,
 )
-from acodex.asyncio.tools.read_thread import (
+from acodex.core.asyncio.tools.read_thread import (
     ReadThreadTool,
     ReadThreadToolInput,
     ReadThreadToolOutput,
 )
-from acodex.asyncio.tools.send_message_to_thread import (
+from acodex.core.asyncio.tools.send_message_to_thread import (
     SendMessageToThreadTool,
     SendMessageToThreadToolInput,
     SendMessageToThreadToolOutput,
 )
-from acodex.asyncio.tools.set_thread_archived import (
+from acodex.core.asyncio.tools.set_thread_archived import (
     SetThreadArchivedTool,
     SetThreadArchivedToolInput,
     SetThreadArchivedToolOutput,
 )
-from acodex.asyncio.tools.set_thread_pinned import (
+from acodex.core.asyncio.tools.set_thread_pinned import (
     SetThreadPinnedTool,
     SetThreadPinnedToolInput,
     SetThreadPinnedToolOutput,
 )
-from acodex.asyncio.tools.set_thread_title import (
+from acodex.core.asyncio.tools import (
     SetThreadTitleTool,
     SetThreadTitleToolInput,
     SetThreadTitleToolOutput,
 )
-from acodex.asyncio.tools.thread_tools import CodexAppThreadTools
+from acodex.core.asyncio.tools.thread_tools import CodexAppThreadTools
 
 ToolCase = tuple[type, type, type, dict[str, Any], str | None, JsonObject, JsonObject, JsonObject]
